Converting accidental use of doubles to qreal in declarative
Declarative had a number of places where double operations were used
instead of qreal, due to the use of double literals. These are now
constructed as qreal literals so that qreal operations are used.

This makes no difference where qreal is double. But on Symbian, qreal
is float, and this give a performance boost for floating point intensive
code.

Task-number: QTBUG-4894

+++ src/declarative/graphicsitems/qdeclarativelistview.cpp
@@ -1457,14 +1457,14 @@ void QDeclarativeListViewPrivate::flick(AxisData &data, qreal minExtent, qreal m
             // the initial flick - estimate boundary
             qreal accel = deceleration;
             qreal v2 = v * v;
-            overshootDist = 0.0;
+            overshootDist = qreal(0.0);
             // + averageSize/4 to encourage moving at least one item in the flick direction
-            qreal dist = v2 / (accel * 2.0) + averageSize/4;
+            qreal dist = v2 / (accel * qreal(2.0)) + averageSize/4;
             if (maxDistance > 0)
                 dist = qMin(dist, maxDistance);
             if (v > 0)
                 dist = -dist;
-            if ((maxDistance > 0.0 && v2 / (2.0f * maxDistance) < accel) || snapMode == QDeclarativeListView::SnapOneItem) {
+            if ((maxDistance > qreal(0.0) && v2 / (2.0f * maxDistance) < accel) || snapMode == QDeclarativeListView::SnapOneItem) {
                 if (snapMode != QDeclarativeListView::SnapOneItem) {
                     qreal distTemp = isRightToLeft() ? -dist : dist;
                     data.flickTarget = -snapPosAt(-(dataValue - highlightStart) + distTemp) + highlightStart;
